refactor: streamline schema handling
