+++ CHANGELOG.md
@@ -59,6 +59,8 @@ Now it:
   dependencies are injected instead of relying on trait properties.
 - All commands and programmatic entry points now build a `Specification` and use
   `GenerationRunner::generateFromSpecification` for generation.
+- `Schema2Class::generateFromSchema` delegates directly to
+  `GenerationRunner::generateFromRequest` without creating temporary files.
 - **Updated tests and fixtures**: Many new test cases covering new functionality such as typed arrays, non-ASCII identifiers, and the programmatic API.
 - **Dependency updates**:
   - New dependencies: `voku/portable-ascii`, `nikic/php-parser`
